creacion de carpeta solicitudScreens  y division de solicitud list para mejor estructura y mas legible
