+++ screens/solicitudscreen/data.js
@@ -1,0 +1,77 @@
+export const estados = [
+  { label: 'Seleccione Estado', value: '' },
+  { label: 'Estado 1', value: 'estado1' },
+  { label: 'Estado 2', value: 'estado2' },
+];
+
+export const municipios = [
+  { label: 'Seleccione Municipio', value: '' },
+  { label: 'Municipio 1', value: 'municipio1' },
+  { label: 'Municipio 2', value: 'municipio2' },
+];
+
+export const parroquias = [
+  { label: 'Seleccione Parroquia', value: '' },
+  { label: 'Parroquia 1', value: 'parroquia1' },
+  { label: 'Parroquia 2', value: 'parroquia2' },
+];
+
+export const bancos = [
+  { label: 'Seleccione Banco', value: '' },
+  { label: 'BDV', value: 'bdv' },
+  { label: 'Banesco', value: 'banesco' },
+];
+
+export const solicitudesMock = [
+  {
+    id: 1,
+    consejo: 'PRUEBAUNO',
+    fecha: '12/05/2025',
+    referencias: [
+      { valor: '1011565854', color: '#20c997' },
+      { valor: '1011584456', color: '#dc3545' }
+    ],
+    total: '990 Bs',
+    estado: 'PENDIENTE'
+  },
+  {
+    id: 2,
+    consejo: 'PRUEBADOS',
+    fecha: '13/05/2025',
+    referencias: [
+      { valor: '2011565854', color: '#20c997' }
+    ],
+    total: '1200 Bs',
+    estado: 'PAGADO'
+  },
+  {
+    id: 3,
+    consejo: 'PRUEBATRES',
+    fecha: '14/05/2025',
+    referencias: [
+      { valor: '3011565854', color: '#20c997' }
+    ],
+    total: '800 Bs',
+    estado: 'PENDIENTE'
+  },
+  {
+    id: 4,
+    consejo: 'PRUEBACUATRO',
+    fecha: '15/05/2025',
+    referencias: [
+      { valor: '4011565854', color: '#dc3545' }
+    ],
+    total: '1500 Bs',
+    estado: 'RECHAZADO'
+  },
+  {
+    id: 5,
+    consejo: 'PRUEBACINCO',
+    fecha: '16/05/2025',
+    referencias: [
+      { valor: '5011565854', color: '#20c997' }
+    ],
+    total: '950 Bs',
+    estado: 'PENDIENTE'
+  }
+];
